requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
aiosqlite==0.20.0
aiohttp==3.11.11
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyJWT==2.9.0
